fix(tier2): clamp duration_ms before int32 cast + use FDateTime::ToString (PR #44)
Two findings on PR #44, both addressed:

P1 / high (Codex + Gemini, both bots converged): cast-before-clamp UB.
  Original: FMath::Min(static_cast<int32>(Raw), kMaxDurationMs)
  When Raw > INT_MAX, the cast-to-int32 invokes undefined behavior --
  the result is unspecified (could overflow to a negative int, wrap
  unpredictably, etc.), and FMath::Min would then compare against
  garbage. Fix: clamp on the WIDE (double) type FIRST, narrow LAST:
    static_cast<int32>(FMath::Min(Raw, static_cast<double>(kMaxDurationMs)))

  This is the same family of bug as PR #39's %g-truncates-precision and
  PR #43's cursor-doesn't-advance-past-inspected-events: silently-wrong
  numeric coercion that passes type-check but mangles values. Future
  numeric-coercion code in this project should follow the discipline of
  clamping on the wide type before narrowing.

Medium (Gemini): use FDateTime::ToString() instead of manual Printf.
  Default format ("%Y.%m.%d-%H.%M.%S") matches the canonical
  YYYY.MM.DD-HH.MM.SS form LogCapture and EventBus emit. Cleaner code,
  same output. Note: LogCapture.cpp + EventBus.cpp still use the
  hand-rolled Printf form (mirrored when I wrote TaskRegistry); a
  follow-up cleanup PR could harmonize them. Out of scope for PR #44.

Tests: 133 still green.

+++ UnrealClaudeMCP/Source/UnrealClaudeMCP/Private/MCP/Handlers/Handler_StartSleepTask.cpp
@@ -83,7 +83,12 @@ public:
                 TEXT("start_sleep_task: invalid_value_shape: 'duration_ms' must be > 0 (got %g)"), Raw);
             return nullptr;
         }
-        const int32 DurationMs = FMath::Min(static_cast<int32>(Raw), kMaxDurationMs);
+        // Clamp on the WIDE (double) type BEFORE narrowing to int32. The
+        // reverse order would invoke undefined behavior for Raw > INT_MAX
+        // (the cast result is unspecified -- could overflow to a negative,
+        // wrap, or worse), and the subsequent FMath::Min would compare
+        // against garbage. Caught by Codex P1 + Gemini high-priority on PR #44.
+        const int32 DurationMs = static_cast<int32>(FMath::Min(Raw, static_cast<double>(kMaxDurationMs)));
 
         // --- register task and spawn worker --------------------------------
